backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pydantic==2.10.3
pydantic[email]==2.10.3
python-multipart==0.0.20
alembic==1.17.2
asyncpg==0.30.0
python-dotenv==1.2.1
